backend/requirements.txt
# Web框架
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0    # PostgreSQL异步驱动
psycopg2-binary==2.9.9  # PostgreSQL同步驱动（用于迁移脚本）

# 数据验证
pydantic==2.12.4
pydantic-settings==2.11.0

# AI服务
openai==2.7.0
anthropic==0.72.0

# 工具库
httpx==0.28.1
python-dotenv==1.0.0
psutil==6.1.1
# MCP官方库（Model Context Protocol Python SDK）
mcp==1.21.0


# NumPy版本锁定（兼容性要求）
numpy==1.26.4

# 向量数据库和Embedding (长期记忆系统)
chromadb==1.3.2


# Transformers（更新到最新稳定版本以修复 FutureWarning）
transformers==4.57.1

# Sentence Transformers（更新到最新稳定版本以修复 FutureWarning）
sentence-transformers==5.1.2

# 中文分词（文风分析）
jieba==0.42.1
